fix: migrated the keepkey passphrase modal to live under wallet context
