+++ src/context/ModalProvider/ModalProvider.tsx
@@ -4,7 +4,6 @@ import { GetStartedModal } from 'plugins/cosmos/components/modals/GetStarted/Get
 import { StakingModal } from 'plugins/cosmos/components/modals/Staking/Staking'
 import React, { useMemo, useReducer } from 'react'
 import { FiatRampsModal } from 'components/Modals/FiatRamps/FiatRamps'
-import { PassphraseModal } from 'components/Modals/KeyManagement/KeepKey/Passphrase'
 import { ReceiveModal } from 'components/Modals/Receive/Receive'
 import { SendModal } from 'components/Modals/Send/Send'
 import { SettingsModal } from 'components/Modals/Settings/Settings'
@@ -15,7 +14,6 @@ import { ModalContext } from './ModalContext'
 // the key is the name returned by the hook and the
 // component is the modal to be rendered
 const MODALS = {
-  keepkeyPassphrase: PassphraseModal,
   receive: ReceiveModal,
   send: SendModal,
   fiatRamps: FiatRampsModal,
